Refactor-fix blank mobile & speed up parallel fetch
